chore: auth 폴더 구조 정리(FSD 구조)
