+++ src/entities/auth/model/user.ts
@@ -1,0 +1,12 @@
+export interface signupUser {
+  email: string;
+  password: string;
+  name: string;
+  companyName: string;
+}
+
+export interface signinUser {
+  email: string;
+  password: string;
+  token?: string;
+}
