--- src/entities/signin/model/user.ts
@@ -1,5 +1,0 @@
-export interface User {
-  email: string;
-  password: string;
-  token?: string;
-}
