--- src/entities/signup/api/index.ts
@@ -1,14 +1,0 @@
-import { User } from '../model/user';
-import { signupApiService } from './service/SignupApiService';
-
-export function useSignup() {
-  const signup = async (userData: User) => {
-    try {
-      const response = await signupApiService.signup(userData);
-      return response.data;
-    } catch (error) {
-      console.error(error);
-    }
-  };
-  return { signup };
-}
