--- src/entities/signup/model/user.ts
@@ -1,6 +1,0 @@
-export interface User {
-  email: string;
-  password: string;
-  name: string;
-  companyName: string;
-}
