+++ src/features/auth/ui/AuthImage.tsx
